Modify error handling

+++ src/main/java/apps/SlideApiResource.java
@@ -1,11 +1,9 @@
 package apps;
 
 import static dev.nklab.jl2.Extentions.$;
-
 import java.io.IOException;
 import java.text.ParseException;
 import java.util.Map;
-
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -16,20 +14,14 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.SecurityContext;
-
-import com.fasterxml.jackson.core.util.DefaultIndenter;
-import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
+import javax.ws.rs.core.Response.Status;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
 import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
-
 import dev.nklab.jl2.web.logging.Logger;
 import dev.nklab.jl2.web.profile.WebTrace;
-import io.quarkus.security.Authenticated;
+import fw.AuthException;
 
 @Path("/api/slide")
 public class SlideApiResource {
@@ -45,39 +37,6 @@ public class SlideApiResource {
     @Inject
     TokenService tokenService;
 
-    // @Provider
-    // public static class ForbiddenExceptionMapper implements ExceptionMapper<RuntimeException> {
-    //     @Override
-    //     public Response toResponse(RuntimeException exception) {
-    //         return Response.status(Response.Status.FORBIDDEN).build();
-    //     }
-    // }
-
-    @GET
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    @WebTrace
-    public Response list(@HeaderParam("x-slide4vr-auth") final String token) throws IOException, ParseException {
-        var userId = tokenService.getUserId(token);
-        System.out.println(userId);
-        System.out.println(tokenService.getToken(userId));
-
-        logger.debug("getList", $("id", userId));
-
-        final var slides = slideService.listSlides(userId);
-        return Response.ok(new ObjectMapper().writeValueAsString(slides)).build();
-    }
-
-    @GET
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    @WebTrace
-    @Path("user")
-    public Response getUser(@HeaderParam("x-slide4vr-auth") final String token) throws IOException, ParseException {
-        var userId = tokenService.getUserId(token);
-        return Response.ok(new ObjectMapper().writeValueAsString(Map.of("userId", userId))).build();
-    }
-
     @GET
     @Path("{id}/{key}")
     @Consumes(MediaType.APPLICATION_JSON)
@@ -90,31 +49,68 @@ public class SlideApiResource {
             format = "";
         }
         switch (format) {
-        case "vcas":
-            final var printer = new DefaultPrettyPrinter();
-            printer.indentArraysWith(DefaultIndenter.SYSTEM_LINEFEED_INSTANCE);
-            json = new ObjectMapper().writer(printer).writeValueAsString(slideService.getSlide4Vcas(id, key));
-            break;
-        default:
-            json = new ObjectMapper().writeValueAsString(slideService.getSlide(id, key));
+            case "vcas":
+                json = new ObjectMapper().writeValueAsString(slideService.getSlide4Vcas(id, key));
+                break;
+            default:
+                json = new ObjectMapper().writeValueAsString(slideService.getSlide(id, key));
         }
 
-        return Response.ok(json).build();
+        if (!json.isEmpty()) {
+            return Response.ok(json).build();
+        } else {
+            return Response.status(Status.NOT_FOUND).entity("Not Found: " + id + "/" + key).build();
+        }
+    }
+
+    @GET
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    @WebTrace
+    public Response list(@HeaderParam("x-slide4vr-auth") final String token)
+            throws IOException, ParseException {
+        if (token == null || token.isBlank()) {
+            throw new AuthException("token is empty");
+        }
+        var userId = tokenService.getUserId(token);
+
+        logger.debug("getList", $("id", userId));
+
+        final var slides = slideService.listSlides(userId);
+        return Response.ok(new ObjectMapper().writeValueAsString(slides)).build();
+    }
+
+    @GET
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    @WebTrace
+    @Path("user")
+    public Response getUser(@HeaderParam("x-slide4vr-auth") final String token)
+            throws IOException, ParseException {
+        if (token == null || token.isBlank()) {
+            throw new AuthException("token is empty");
+        }
+        var userId = tokenService.getUserId(token);
+        return Response.ok(new ObjectMapper().writeValueAsString(Map.of("userId", userId))).build();
     }
 
     @POST
-    @Consumes({ MediaType.MULTIPART_FORM_DATA })
+    @Consumes({MediaType.MULTIPART_FORM_DATA})
     @Produces(MediaType.APPLICATION_JSON)
     @WebTrace
-    public Response create(@HeaderParam("x-slide4vr-auth") final String token, final MultipartFormDataInput form)
-            throws IOException {
-        final var slide = new SlideFormBean(form.getFormDataMap());
+    public Response create(@HeaderParam("x-slide4vr-auth") final String token,
+            final MultipartFormDataInput form) throws IOException {
+        var slideForm = new SlideFormBean(form.getFormDataMap());
+        if (token == null || token.isBlank()) {
+            throw new AuthException("token is empty");
+        }
+        var userId = tokenService.getUserId(token);
+        var key = slideService.create(userId, slideForm);
+        pptx2pngService.request(userId, key, slideForm.getContentType(), slideForm.getExtention());
 
-        final var userId = tokenService.getUserId(token);
-        final var key = slideService.create(userId, slide);
-        pptx2pngService.request(userId, key, slide.getContentType(), slide.getExtention());
-
-        return Response.ok(String.format("{messag   e:'%s', data-size:'%d'}", slide.getTitle(), slide.getSlide().length))
+        return Response.ok(new ObjectMapper()
+                .writeValueAsString(Map.of("title", slideForm.getTitle(), "length",
+                        slideForm.getSlide().length, "path", "/api/slide/" + userId + "/" + key)))
                 .build();
     }
 
@@ -123,8 +119,12 @@ public class SlideApiResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @WebTrace
-    public Response delete(@HeaderParam("x-slide4vr-auth") final String token, @PathParam("key") final String key) throws IOException {
-        final var id = tokenService.getUserId(token);
+    public Response delete(@HeaderParam("x-slide4vr-auth") final String token,
+            @PathParam("key") final String key) throws IOException {
+        if (token == null || token.isBlank()) {
+            throw new AuthException("token is empty");
+        }
+        var id = tokenService.getUserId(token);
         var result = slideService.delete(id, key);
 
         return Response.ok(String.format("{delete:'%s', status: %s}", key, result)).build();
